extend lib/ui keypress listener and render from app_base
